+++ src/components/settings/settings-form/settings-form.js
@@ -8,6 +8,7 @@ import {
   PaneFooter,
   Button,
 } from '@folio/stripes/components';
+import Toaster from '../../toaster';
 
 export default class SettingsForm extends Component {
   static propTypes = {
@@ -20,6 +21,8 @@ export default class SettingsForm extends Component {
       updateIsPending: PropTypes.bool,
     }).isRequired,
     title: PropTypes.node,
+    toastAction: PropTypes.func,
+    toasts: PropTypes.array.isRequired,
   };
 
   renderFooter() {
@@ -69,6 +72,8 @@ export default class SettingsForm extends Component {
       children,
       title,
       formState,
+      toasts,
+      toastAction,
       ...formProps
     } = this.props;
 
@@ -89,6 +94,12 @@ export default class SettingsForm extends Component {
           )}
           footer={this.renderFooter()}
         >
+          <Toaster
+            position='bottom'
+            toasts={toasts}
+            toastAction={toastAction}
+          />
+
           {children}
         </Pane>
       </form>