+++ src/components/toaster/toaster.js
@@ -19,6 +19,7 @@ class Toaster extends Component {
   static propTypes = {
     // the position in which toasts are displayed
     position: PropTypes.string,
+    toastAction: PropTypes.func,
     // array of toast messages to display
     toasts: PropTypes.arrayOf(PropTypes.shape({
       // the type of toast: warn, error, or success
@@ -37,7 +38,8 @@ class Toaster extends Component {
   };
 
   static defaultProps = {
-    position: 'bottom'
+    toastAction: () => null,
+    position: 'bottom',
   };
 
   constructor(props) {
@@ -52,6 +54,12 @@ class Toaster extends Component {
     return nextProps.toasts ?
       { toasts: nextProps.toasts } :
       null;
+  }
+
+  componentDidUpdate() {
+    if (this.state.toasts.some(toast => toast.type === 'success')) {
+      this.props.toastAction();
+    }
   }
 
   destroyToast = (toastId) => {