+++ src/redux/actions/update-custom-labels-success.js
@@ -1,0 +1,7 @@
+export const UPDATE_CUSTOM_LABELS_SUCCESS = 'UPDATE_CUSTOM_LABELS_SUCCESS';
+
+export function updateCustomLabelsSuccess() {
+  return {
+    type: UPDATE_CUSTOM_LABELS_SUCCESS,
+  };
+}