+++ src/redux/actions/update-custom-labels.js
@@ -1,0 +1,8 @@
+export const UPDATE_CUSTOM_LABELS = 'UPDATE_CUSTOM_LABELS';
+
+export function updateCustomLabels(payload) {
+  return {
+    type: UPDATE_CUSTOM_LABELS,
+    payload,
+  };
+}