+++ src/redux/epics/index.js
@@ -2,3 +2,4 @@ export { default as updateEntityTags } from './update-entity-tags';
 export { default as createGetAgreementsEpic } from './get-agreements';
 export { default as createAttachAgreementEpic } from './attach-agreement';
 export { default as createGetCustomLabelsEpic } from './get-custom-labels';
+export { default as createUpdateCustomLabelsEpic } from './update-custom-labels';